Move filtering code into own package
Going to need to add some more modules, makes sense to make it non-data
product specific
